Split testing pkg into separate submodules
Instead of keeping things in __init__.py put them in separate submodules for readability.
